feat: Separates Drawer and ResizableDrawer
